The next round of the ECIU implementation.
The previous commit worked, but was only single-threaded.  This enables
multiprocessing, but yet only for a single box.

Solving the ECIU set of a stochastic problem is currently implemented as
a recursive algorithm.  Stage 1 is solved for each assumption, and then
stage 2 is solved for each prior assumption and each assumption.  Then
stage 3, and so on.  This is currently a naïve implementation that uses
parallelism, and is a rough approximation of a breadth-first search.

Of note, the parallelism is somewhat poorly handled in the script,
mainly because I need to learn some of the esoterics of the Python API.
The number of concurrent jobs allowed is 1.5 times the number of cores.
So for Cygnus, where each node has 8 processors, this script allows 12
concurrent solves.

I believe this implementation to be portable to Windows, though I do not
know how comparatively efficient Windows would be given the expense of
creating new processes.

Next steps of implementation:
  - Allow master to be a "passthrough" to disk.  Currently, the master
    process stores the variables for each solve in memory.  This gets
    unwieldy fairly quickly, and at only 5 stages for the simple Temoa
    Island, our 16 GiB is not enough.  One step to solving this would be
    to convert the algorithm into a depth-first search.  Then, only the
    minimal slice of solutions would need to be actively present in
    memory at any point in time.

  - Checkpointing.  Given long running iterations, it is more than
    conceivable that something could break mid-solve.  Thus, having
    checkpoints automatically enabled would be a very nice addition,
    especially in preparation for the next item:

  - Enable computation on the entire cluster.  As noted above, the
    current implementation is limited to a single machine.  We have a
    cluster, and next need to use it.  I imagine this would be
    done by sub-masters on each node, and passing information back to
    the "passthrough" master instance.