Provide a cleaner error message.
For the first (and simple) case of a typo or misunderstanding of the
integer-only requirement of the time_exist and time_future sets, this
commit moves the integerness check to our validation routines.

A number of folks have noted to me that errors produce "impenetrable
output".  This is because most Temoa errors currently come with a long
traceback.  For coopr_python invocations, if the precise error can be
determined, a very specific error message doesn't need the traceback.