Unify temoa_stochastic with temoa_rules.
Since the only difference is the objective function, and the only
difference there is the fact that the stochastic version needs it
separated out along the stochastic set (time_optimize, in this case),
this commit makes use of the previously created PeriodCost_rule.

Also, remove the annoying warning about overriding a "block".  It is
indeed what we are doing, so explicitly remove the original objective
function.