Implementation of growth cap and seed.
The GrowthRateMax parameter allows modelers to specify a percentage cap
on the amount of year over year growth of a technology class.  The
GrowthRateSeed specifies the amount of capacity in excess of the growth
rate that the model may build.  The equation currently implemented is:

CapacityAvailable[p,t] <= GrowthRateMax[t] * (CapacityAvailable[p-1,t] + GrowthRateSeed[t])