Update GrowthRateConstraint and TI_2013
The growth rate constraint would previously not work for technologies
that did not have all vintages defined.  This iteration works for
vintages in time_optimize missing as long as there is no gap.  For
instance if,

time_optimize := 2010  2015  2020  2025  2030  2035  2040 ;

and e_dummy is defined for 2015 through 2035, continuous, then it works.
However, if e_dummy has a "hole", for example if defined for 2015, 2020,
skips 2025, and then 2030 through 2040, then this constraint is still
broken.  I hope that case doesn't become an issue.