Assorted calibrations.
This is still a toy model, but one of a little more intricacy that what
we presented at IEW last year.  We've added some more technologies,
added existing capacity (ECAP or ExistingCapacity), and updated prices.

As we go, we're starting to add units to everything, but I continue to
wonder if we couldn't automate this process.  Certainly at the next
level (e.g., a GUI and DB), but model checking is important to do at
many levels.

One change that we made -- but forgot to /actually/ check if it made a
difference, so I mention it here as some form of reminder -- was to move
the emissions accounting from purely on the import of fuel and negative
emissions for use of a technology (e_ngcc_seq_s in this case), to no
emissions on import, but a more distributed set of emission costs based
on individual process use.