Next round of ECIU implementation.
We not yet to multinode calculation, but this commit brings an
efficiency gain.  The win here is in using the file system and pickling
of information to pass data between processes.  This enables the master
node not to grow in size, and makes ECIU calculations with 5+ stages
feasible from a memory perspective.  However, this is still a breadth
first algorithm, so not a win.

The other "big" win is no longer performing any redundant solves.
Previously, we solved every node for every possible combination.  This
was algorithmically correct, but redundant.  Suppose, for example, a
binary ECIU tree of 2 stages:

     ------ Outcome 1
    /
Root
    \
     ------ Outcome 2

At the root node, we clearly need to solve the perfect foresight of
Outcome 1 and Outcome 2 (two solves).  At Outcome 1, however, we don't
need to resolve the Root-Outcome 1.  We just did!  However, we /do/ need
to solve at Outcome 1 the realization of Outcome 1 with the previous
expectation of Outcome 2 (one solve).  A similar argument goes for the
number of solves at the Outcome 2 node (one solve).  In the previous
iteration, we would have performed 6 solves.  Now we only perform 4.  It
gets better for more stages.

Also, it's currently untested, but the ability is there to restart a
computation.  This is a nice side effect of the current implentation.
I hadn't originally planned it, but it worked out that it was easy to
implement.  I don't know that it will stay in future iterations, but I
note it here.

I implemented a couple of other tidbits along the way:

 * There is no need for the ECIU calculation to have normal priority.
   In fact, Temoa should never need full priority, so it now nices
   itself to the lowest priority.  The win on desktops is that end-user
   computers should hopefully stay more responsive, and this also plays
   nicer as a good "cluster-izen" for other users. (os_nice ...)

 * Realized a need for smart handling of exceptions in child processes.
   We now pass exceptions (as complete as necessary for our purposes) to
   the master process, for it to stop computation, before re-raising
   the exception.  It's not 100% ideal, but it's the best I currently
   know how to do in a multiprocessing environment.

 * Implemented another Temoa specific error: TemoaInfeasibleError

 * If the user has installed the Python packages 'setproctitle', then
   the ECIU calculation changes the executable name of the child processes
   to inform anyone watching of how many solves are yet to go in each
   stage.  Helpful if one is watching, through, for example, htop.  What
   I need to do is document this so that folks can know to install that
   package ...