Update documentation
Since my thesis, we have made a few changes and additions.  For
instance, we have some new parameters (e.g., DemandDefaultDistribution,
DemandSpecificDistribution).  We accordingly changed the implementation
a bit that required an update to the documentation to follow the
explanation of how to make a constraint.

I also made a number of editing changes since I wrote this document
while writing my thesis.  I incorporated most of them.