Dynamically choose where to get jQuery.
As we are not a jQuery project, just a hapless application using it,
let's not store it in our repository.  Instead, utilize a CDN (content
delivery network).  The general idea with a CDN is to take advantage of
the fact that these files are used on *many* sites; if a user has
already downloaded jQuery for another site, there's no need to
redownload it.  This saves the user's download time, the user's
bandwidth, our bandwidth, and generally saves the net some transferred
bytes.  Win-Win.

Meanwhile, while developing, the page will be reloaded thousands of
times.  In that case, it behooves everyone for the developer to have a
local copy.  The JQUERYCD variable in the settings files enables this.
For production, simply do not set it in the user_host.py settings file.

With the .gitignore addition, automate ignoring of jQuery files.