Complete refactor to single process form wrapper
The first round of implementation was using ad-hoc $.post actions based
on as-I-went coding.  Due to my use of the jquery-editable plugin, this
included (effectively) an ad-hoc form for every edited element.  This
new implementation pre-wraps -- if the user is logged in -- every
process table in a form.  Thus, there is now a single function for
submitting new information, but the POST data changes for the current
parameter.  This is not a complete implementation because the interface
does not currently account for if a user changes fields, in which case
there would be multiple parameters sent to the update view.  It's a work
in progress.

Along the way, I also:

 * Made the 'showStatus' function slightly more flexible, allowing the
   caller to specify what kind of message it's displaying.  Currently
   have only implemented 'info' and 'error', with the 'error' being the
   default.

 * Have also implemented the ability to actually remove non-tabular
   parameters from the model: if the user completely empties the field
   before submitting, the UI will interpret that as "delete this
   parameter from this process."

 * For items that are deleted, The UI now reloads the process list.
   This is a Q&D method of putting the UI in a known state, allowing me
   to not have to worry about cascaded removals (e.g., removing an
   efficiency will also remove all associated emissions, removing a
   process discount rate replaces the '(GDR)' in the field).  It's not
   the most elegant or visually pleasing method, but it is simple.