Add -V or --version command line arguments.
At some point in the near future, I suspect a user will ask for help and
the version of Temoa in use may be a pertinent detail.  We've already
partially had this problem with Fan.

In a follow up commit, I'll have the deploy script auto change the
TEMOA_* variables when creating the temoa.py ZIP file.