Update documentation: links, patches
Update to the text of the patches section.  We now only accept AGPLv3+,
not just AGPLv3.  The plus sign indicates that if the Temoa Project
accepts any outside contributor's code, we are limited to the AGPL but
are free to use any later version (at our discretion) than v3.

A couple of links were dead, and added a couple pursuant the patches
section update.