Refactoring of JS code.
process_interface was a first attempt and it was showing.  I've
refactored a large part of the JS code and will, in the next few commits
include the requisite backend code.

The biggest pieces are that TemoaLib.js (formerly process_interface.js)
does a lot less DOM manipulation, leaving the building of Forms to
the webserver.  The original idea of passing data back and forth through
JSON wasn't wrong, but required some reinventing of the wheel.  Django
has a lot of niceties (like Form validation) that we can now take care
of on the server.

Also, for speed of development, have added a jQuery Hotkeys plugin and
subsequently added a shortcut "Ctrl+Space" which reloads the Javascript
and CSS in place.  It may seem small, but it reduces a common 2-3 second
operation to less than a quarter second.  Due the number of reloads
necessary for development, this adds up quickly.