Major Refactor.
I bit the bullet and have implemented a form for every DB row.  This is
more in line with how Django "thinks" and has reduced the code bloat
tremendously.  The expense is slightly more complexity of the DOM, but
who cares.  We're starting to settle on an API, and I can document it at
some point in the hopefully near future.