Dynamically change form input type.
The choices input is nice because it means I don't have to write my own
validation routines.  Unfortunately, it is not a natural overlap with
the TextInput.  So, I've opted to split the difference by converting a
forced-choice option to a readonly text field.  Looks a lot nicer, and
still lets the user see the options.