Refactor and inform Javascript.
I'm no longer using the the <prefix>form variable name.  The form is
just to be called 'form'.  Correcting that for this efficiency function.

Also, if there is a form error then we let javascript know through use
of the 422 status "Unprocessable Entity" status code, and put the error
messages in the request context.