Update cookie handling.
TemoaDB now uses a session cookie to store the state of the UI.
Previously, the UI would do this on it's own.  However, as I'm migrating
to passing back pre-rendered HTML instead of JSON, I need a different
method to pass back "meta" information.  Cookies to save the day.

Very simplistic information, only:

 * analysis_id
 * process_ids
 * username

This is a one-way passing of information, as the server does not use
this cookie for any decisions it makes.  To strengthen this point, the
UI removes this cookie after processing and before the next request.