Always reattach listeners.
As part of the refactor from JSON to HTML from the server, the UI now
needs to attach listeners to the new elements.  The problem is that the
UI replaces parts of the current DOM with the updates from the server,
the event functions also get removed.  This replaces the listener on the
analysis change list.