Add and remove processes.
This accomplishes the addition and removal of process through the older
"blue plus" and "red minus with shift" functionality.

There was some work involved in finagling with Django's
ModelChoiceFields innards.  The ModelChoiceIterator assumes a queryset,
but in this case we need a strict subset: the final period of the
in the Analysis is not a valid period.  The standard Python slicing
operation ([:-1]) understandably does not work with querysets (because
the underlying SQL doesn't have a notion of "all but the last row").
So, two new Form classes.  Perhaps I'll add these to Django Snippets at
some point.

Also, removed some unintentially committed debugging messages.