Cleanup UI Events.
While transitioning between HTML and JSON for information passing, the
HTML reload was letting events "pile" up on action elements.  This
cleanup ensures that all events are removed before new ones are added.

Added a .cancel class for DOM manipulation and differentiation between
an already added row and a row that has yet to be added.