More error checking on model.
Though I don't technically need to check that passed value is in the
vintage table, this way allows a better error message.

The bigger issue is a check against the last vintage, which is not a
valid option.

At some point I need to add these checks to DB triggers, but for now,
this works.