Enable short-lived EJS templates.
The problem with client-side templates is that if an update happens, the
browser may not get notified for sometime, depending on when the cache
finally goes stale.  If not specified, automatic cache decisions can
last for a week.  This change should enable short-lived caching of only
5 minutes, which should not be a problem since these templates are
loaded only once when CanJS first uses them.