Clean up saving Vintages.
If a modeler removed the final year in vintages, then the frontend
correctly did not treat the next final year as a valid vintage, but the
data was still lying around in the DB.  Though this behavior wouldn't or
shouldn't prove to be a problem, it also not expected.  Now, all
associated data is correctly dropped from the DB when a vintage is
removed.