The advice is: don't force on other scripts.
"use strict" is a nice idiom, but the reality is that not all scripts
may use it.  We should be fine since I've been developing with it in a
global context, but who knows what library I (or someone) might decide
is important to utilize, and that does not follow work under stricter
browser enforcement.