On error, focus cursor automatically.
I believe that it doesn't "feel" correct for all fields to lose focus
when a form is submitted.  However, we disable the form inputs to
approximate some semblence of atomicity ("Hey user, with your 1 KB/s
connection: don't make a change until we've verified we're in sync with
the server!").  The disabling makes them lose focus, so when we reenable
them, give one of the problem inputs (if any) the focus automatically.