CanJS updated terminology.
The CanJS devs finally realized that what they had been calling Observe,
most of the rest of the dev world would have called a Map.  There was a
certain logic to Observe in that it "observes" it's data, but that's to
be expected given outside knowledge of the CanJS library.  So, a simple
rename.