Rebuild Parameter div upon showing.
There is an order of operations bug that prevents updates if the
Parameters Control is always around, and other updates happen.  There
is perhaps a cleaner solution, but this one is quick: when the div is
closed, remove the Control.  When the user requests the Control by
clicking Analysis Parameters, then dynamically build the div again.

This results in a slight delay when the user clicks it, but hopefully
won't be too noticeable.