Be slightly more judicious in what we accept.
Specifically, the URL can only contain alphanumeric characters or an
underscore and must be followed by '.ejs'.

This URL gets turned directly into a filename -- that we
indiscriminantly serve directly -- so, we need to ensure that there are
no directory characters (like '/'), and will also need to ensure that
the directory of client_templates contains no symbolic or other links to
sensitive information.