Update per save_to_server code.
This was the initial implementation of "save_to_server".  However, I
altered the logic slightly when I incorporated the save_to_server with
the other models.  However, this is effectively now C&P code, so just
correct the logic, until I have a chance to properly reduce the code per
DRY principles.