Change the way Del buttons are enabled.
Previously, one had to push 'shift' to even see the buttons.  It's a
slick approach to prevent mistaken deletes on this platform where there
is (currently) no undo.  On the other hand, it's not very discoverable.
This new approach draws the buttons, but disables them.  The title
attribute then says 'Press shift to enable.'.  This has the benefit of
being a bit more discoverable, while still deterring mistaken DELETE
commands to the server.