Development function: inform me of error.
This function, "Shift+Space" downloads the file QuickFunction.js from
the server and executes it dynamically in place.  This is _very_ useful
for development because it means the whole page does not have to reload,
just the sub function on which I'm working.  Mucho faster.  However, if
I had a typo, all I would get is 'Error'.  This helps narrow down what's
gone wrong.