Remove "Reporting Variables" from solution.
Fixes #13.  Sort of.

I suppose I could still incorporate this script automatically into a
post-process section, but for now, it's required to manually execute it.
For example:

    $ ./temoa.py  utopia.dat  |  ./temoa.py  --calculate_report_variables -
       # or
    $ ./temoa.py  utopia.dat > utopia.sol
    $ ./temoa.py --calculate_report_variables utopia.sol > utopia_extended.sol

So, still not ideal, and as #13 states, could lead to a mismatch, but
better than nothing as larger problems will not run much faster in the
Pyomo stage.  "Don't let 'perfect' be the enemy of 'good'."

The astute observer will note that some previously implemented pieces
are no longer implemented, like InvestmentByTech.  That's pending as
round two of this needs to pull in certain details from parameters.