Another ROOT_URL conversion.
This is not the most elegant, as I'm yet again tying the model to a URL
scheme.  I need to do some research as to the best way to interface
between the browser and the server via the browser model.  Ideally,
there's a method as clean as Django's 'urls.py' model.  At some point,
this code could do with some targetted refactoring.  However, I'm
currently in the mode of "get it working" and not letting "perfect be
the enemy of the good".