Attempt to fail gracefully on SIGPIPE.
Fixes #17.

This is actually a partial fix, but due to Coopr's implementation, a
complete fix is not possible at this juncture.  This fix handles a
common case where Temoa's stdout pipe is closed before it can finish
writing.  (This might happen if a user interactively pipes the output to
a pager, and then quits the pager before reading all the data from
Temoa.)