Fix bug with fixed variables not shown in output.
The addition of --fix_variables showed a problem with the results of
Temoa: namely that fixed variables were not shown.  This fix is more
explicit in nature than the previous general approach to printing
whatever was returned by the solver.  Now, this is Temoa specific, but
it has been all along.  I just hadn't put that together the first few
go-rounds.