Do not select already disabled buttons to disable.
When initially disabling the form element, it doesn't matter if we
disable already disabled buttons.  But, after the interaction ends
(saved or failure), they were part of the group that was enabled again.
For the disabled buttons for which we request the user press shift, this
does not make sense.  So, don't select them initially.