Only set fields that user has changed.
Fixes #19

Some faulty logic initially that equated "user didn't send this piece of
data" with "user wants to remove this piece of data".  That's patently
not what we want, as we only want the user interface to send data that
has actually changed.  Easy fix: if user didn't send those pieces of
information, then remove the field that Django processes.