First round of TemoaDB tests.
This begins testing of TemoaDB.  So far, only 2 models are checked.
In addition to testing all the models, we also need to test the API
interactions.  (i.e., the GET and POST requests from authenticated and
unauthenticated sources.)

Run the tests with manage.py:

    $ python manage.py test

Note that 'coverage' is a handy tool for visually seeing which lines
have actually been executed.  Run like this:

    $ coverage run --source='.' manage.py test
      # This will write a file, usually '.coverage' to the directory

    $ coverage report    # text/terminal output
    $ coverage html      # creates directory 'htmlcov'.