Squash test-identified bug.
Period_0 and Vintages go hand in hand.  Period_0 should not be empty,
and period_0 must be a member of the analysis vintages.  However,
because the analysis must be created first, Period_0 cannot be a foreign
key into the Vintages table.  Thus, this needs to be checked by the
application code, which I forgot to write, initially.  Done.