Remove empty fields.
These might occur if a modeler accidentally types a comma twice while
putting in vintages: 1,2,,3,4.

This was a forgotten addition originally, and without it, the following
if-check for no vintages is always false.  This wasn't caught because of
the front-end layer, but now it's fixed in the backend.  Test for this
following shortly.