Correctly convert to an integer on save.
If someone were to instantiate an Analysis in a REPL or in code, and for
some reason set period_0 or vintage to a non-integer, the .save() call
is not enough to convert that instance's period_0 or vintage to an
integer.  New copies obtained from the DB will correctly be integers,
but not this original.  For example, /before/ this commit, this might
happen:

    >>> a = Analysis()
    >>> a.name = 'Some Name'
    >>> a.description = 'Some description'
    >>> a.global_discount_rate = '0.05'
    >>> a.period_0 = '1.4'
    >>> a.save()
    >>> a.period_0
    1.4

!!! That should be an integer!  Since the DB just silently converts it
to an integer, things are mostly hunky dory, but this view is now out of
date:

    >>> b = Analysis.objects.get( pk=a.pk )
    >>> b.period_0
    1

Clearly the DB saves it correctly.  After this patch, the value for
period_0 and .vintage now update on save:

    >>> a.period_0
    1

Phew!  Not a big bug, but hopefully preventing a hapless command liner
from themselves ...