Django already provides clean_fields.  Use it.
Bah, it appears that I missed the fact that Django already provides
self.clean_fields(), which does exactly I was trying to do.  Django does
not, however, automatically call it.  On their end it's a good
generalist decision, but on our end, we want it to be called.  So, the
new test now is to ensure that it is, indeed, called.