Fixes #34 "Update Process"
Background: TemoaDB attempts to only communicate between the server and
webclient what is necessary when it is necessary.  The client was
correctly only sending the loan lifetime value, but the server was
updating all parameters.

The one line fix was an iterator oversight within Python: one should
never update an object while using a related iterator.  A simple list()
fixes the issue.