A simple CSV converter for Temoa's default output.
This script takes Temoa's index-value-per-line format and converts it to
CSV format.  I do not know of a way to pre-guess what orientation folks
will need the data, so the table rows and columns chosen are arbitrary.
I hope the function Act_2CSV well-enough explains how it works, so that
others may manipulate this script per their needs.