Simplify LP file logic.
From the modeler's perspective, I believe the only reason they'll want
to look at an LP file is for edification or to make sure the model is
doing what they think it's doing.  There's thus no need to give the
modeler the option to use symbolic solver labels, which is a gritty
detail of Coopr.  Thus, hide that fact: if the user requests to see the
LP files, use symbolic labels.  Full stop.