Fix floating point issue.
The old iteration of the algorithm attempted to calculate the number of
characters in the fractional portion of a number with this idiom:

    numChars = len(str(abs(val - int(val))))

Unfortunately, for something like 1.000000000004, this results in the
fractional part being:

    >>> 1.000000000004 - 1
    3.999911513119514e-12

This is correct for the numChars, but not correct when we later print it
out, because  1.000000000004 turns into 1.999911513119514e-12, which is
not correct.  This new algorithm splits on the decimal of the entire
string, which should correct this bug.