pformat: str -> repr
In terms of data correctness, str() uses up to 12 decimals of precision,
which perhaps reads nicer but does not produce round-trippable values.
That is,

    >>> float(str(x)) == x

is not guaranteed to always return True.  However,

    >>> float(repr(x)) == x

is guaranteed to always be True.

The few extra digits of precision are of vanishingly small increased
utility, but it's a correctness issue.

Note that if we ever get to Python 3.4+, this should be a non-issue.