Do not alter function argument.
The previous iteration of this function altered the input by converting
each element of the passed argument into a string.  This update now does
the necessary string preparations in place, with no surprising
side-effects.  This shouldn't have been a problem, but from the
department of "defensive programming" ...  This implementation should be
slightly more efficient too, not that anyone would notice as this
function is _not_ the bottleneck.

Also, update function comment.