Update tests to work with new technology assumption.
The technology assumption is that techs are attached to an analysis.
Period.  This is a change where technologies were intended to be
"database wide" so that users could share data.  However, since DB
technologies had only C2A, it was rather pointless.  Now technologies
will be directly tied to the analysis, and to share access, users will
need to copy entire analysis to their namespace before changing.

Of course, that copying functionality is not yet implemented ...