Robustify column alignment code.
The previous code erroneously assumed Coopr guaranteed a float value.
Thanks to a forum report, I've now learned that Coopr's solver parsing
routines are not the same.  In particular, eval() and float() are two
different methods to ascertain a value from input.  In short, the CBC
solution parser uses eval(), which returns either an integer or a float
value in this case, while the GLPK, CPLEX, and GUROBI parsers use
float().  The float() function is the correct method to use here, not
eval(); while this is technically a bug in the CBC parser -- or at the
very least a weak assumption -- it's no problem to robustify Temoa's
code.