A first set of JS/UI tests.  Well, test (singular).
This is a single test of getCookie().  I chose it for it's simplicity
while I got the rest of the testing framework in place.  To make the
getCookie() function accessible, I'm experimenting with a Temoa
namespace in Javascript.  This will give unit-test access to various
functions.  The associated refactor is not terribly invasive, just
renaming use of getCookie() to Temoa.fn.getCookie().  This commit also
starts the unit_test_ui.html page, which will hopefully be a one-stop
shop to check to see if an instance is running correctly.  Further
commits to come.