A view unit test.
This is more what tests /should/ look like as I'd like to test /our/
code, not the rest of the framework I didn't write (i.e., Django).  I'll
have to update the other tests in this file accordingly.  (I must be
learning!).

This begins to highlight some of the changes I missed in regards to the
Technology/Analysis rethink.  Or, this is a follow-up to

    2c5ec7b0ca6d5515ec4d9b1bc48ba3df278f54ad
    "Update tests to work with new technology assumption."