Remove non-functional code.
This code hasn't worked.  Ever.  I had intended it to be able to debug
EJS templates, but I was still learning the necessary Javascript at the
time and was hoping it would "just work".  Meanwhile, it requires use of
an outside EJS debugging script, which I was unsuccessful in finding on
line.

This may become moot, however, as I'm looking at Mustache templates as a
possible replacement.  The syntax looks to be much cleaner, which should
make building client-side templates a lot easier.