Dynamically load TemoaDB_functions.js.
This is perhaps not strictly necessary, but doing it this way enables
the entry point of TemoaUI to "gracefully" fail, at least giving the
user /something/ if an error occurs.  This is in opposition to what
would happen if these were loaded as part of the root HTML document
(<script href=...) tags, where the browser would just sit there not
providing much information.

Also, make BeginTemoaDBApp() utilize the Temoa.fn namespace, where
appropriate, and include a jQuery function (borrowed from their
website).