Move library loading into TemoaDB.
I like doing it this way because now the user gets a message if a
required library doesn't load.  Or, at the very least, this is more
programmatically flexible.

Also, there's really no need for the CDN thought process for our
project: we are _so_ small.  We deliver all we need, and that's a good
thing because then we can be absolutely sure of the quirks and versions.