Begin the transition to mustache templates.
Mustache templates are much cleaner (e.g., simpler syntax), and by
virtue of their logicless nature, enforce critical logic handling at a
more appropriate place.

This first commit on this path also enforces mustache only templates
(which is important going forward because CanJS has deprecated EJS
templates) at the url level.

The changes in views.py are necessary because otherwise Django attempts
to interpret these templates prior to sending them to the browser.  The
right answer is of course to make these static files, but that
reintroduces the original reason I implemented this non-static pull:
annoyingly long caching by the browser of these client templates.  So
until I figure out a better way to invalidate the browser cache
efficiently, this is a development benefit for me.  Meanwhile, we're a
small, small, small project, so I can get away with this.