Change interaction to a matrix-style input.
A long way to go yet, but this is a big step in changing the interaction
from a per-process input to a technology-matrix style input.  A big
hurdle was in figuring out the non-sparse representation of the data for
the (almost) logic-less Mustache template.

Note the move of the analysis_technology_detail.ejs to
TechnologyDetail.mustache.