Refactor ProcessForm
As we're now Tech-matrix oriented on the UI end, do some cleanup on the
ProcessForm functionality.

Also, and perhaps more importantly, with the realization that all
analyses will *not* be functionally correct at some point while users
make changes, stop trying to be so correct.  For example, allow a user
to specify existing capacity for a non-existing vintage process.  Now,
we'll do the model correctness check in only one or two places.  This is
a much saner approach, and also means a lot of the (in hindsight) ad hoc
"correctness" checking can be removed.