Incorporate the "outside functions"
Charlie had originally had the activity function as part of an outside
function.  This may be the path to take to enable modeler-specified
iteration functions but I haven't thought as deeply about it as he.
Meanwhile, this implementation is a bit faster as there is no need for
parsing of the Temoa output files.

The biggest piece to figure out, however, was how to dynamically add the
next constraint or objective function.  It turns it's easy unless you
follow the Coopr/Pyomo coding guidelines: it overwrites the function
you've just written if another function with the same name exists in
another module.  That was a fun one to debug.