Update Temoa copyright and license to GPLv2
This update brings Temoa into compliance with the NC State Office
of Technology Transfer. Temoa is now licensed under GPLv2 and the
copyright belongs to NC State University. In making this update,
I considered a more permissive license, such as BSD, in order to
provide more flexibility to future users. However, in the end,
I believe that the model should remain in the public domain, esp.
if it is used to inform public policy. Hence the choice of a
copyleft license.