Update TechInput/OutputSplit_Constraint
The previous version of TechInputSplit_Constraint and TechOutputSplit_Constraint had two limitations: (1) commodity shares were forced to sum to 1, and (2) the shares remained fixed through time. This commit addresses both of these shortcomings. The parameters TechInputSplit and TechOutputSplit are now indexed by model time period, and do not need to sum to 1. When the sum does not equal 1, the model optimizes the unspecified shares.