Remove Temoa-specific error classes
Several error classes were defined in temoa_common.py. They were not required as they did not perform any model-specific error handling. For simplicity, we simply raise a general 'exception' and provide a descriptive error message. In addition, temoa_common.py was also removed as it is now not needed. The function 'get_str_padding' moved to temoa_initialize.py and the function 'iter_in_chunks' was not being used.