Modify utopia case to ensure that database and DAT files produce same results
Previously, results produced with temoa_utopia.sqlite and utopia-15.dat did not match. There were two issues. First, the existing capacity of RHO was mis-specified in the database version. Second, some parameters were specified as integers in the DAT file, and in Python 2.7, this causes a problem when integers are divided (e.g., 1/2 = 0 instead of 0.5). Latter problem solved with an import statement at the top of temoa_rules.py.