Complete major refactor of temoa_lib.py -- now temoa_run.py
Functions within temoa_lib were refactored to streamline the code and functions were added for redundant operations. Functions to initialize and check the model were offloaded to a new module temoa_initialize in a previous commit. Thus, the name of temoa_lib.py is changed to temoa_run.py, since this module now deals with constructing an instance, solving it, and handling files. Also note that MGA-specific functions moved to a separate module called temoa_mga.