Fix lex parser to make compatible with Python 3
When Python 3 reads a file, it returns a list of integers if the 'b' flag is there, otherwise it returns a string. The lex parser expects a string. So if 'b' is there, it will raise an error. In Python 2, 'rb' does the same thing as 'r', which always returns a string.