Adjust code so storage duration specificied as fraction of a year
Previously, units were hard-wired into the code. The storage duration was specified as a number of hours, which when multiplied by the capacity variable, creates energy units that are not necessarily consistent with the energy units for that process. The kluge had been to hard-wire the units in the code, assuming activity was in PJ. Specifying storage duration as a fraction of the year and using the CapacitytoActivity parameter fixes the issue. As part of this change, the storage unit is initialized at full charge.