Add new V_Curtailment variable to track curtailment of non-dispatchable renewables
This curtailment variable was added to track curtailment of renewables. It was motivated by the need to model a constraint in scenarios where a renewable target as a function of final energy demand is specified. Without this revised formulation, the model can count curtailed energy toward the target. In order to use this new functionality, technologies that can be curtailed (e.g., wind and solar) must be added to the tech_curtailment set. Additional changes to constraints were required. For technologies in tech_curtailment, the Capacity_Constraint and ProcessBalance_Constraint include an explicit curtailment variable. In addition, the CommodityBalance_Constraint and Demand_Constraint are changed to strict equalities in order to avoid overproduction.