Add NEOS solve functionality to config file
With this commit, users can now use the NEOS server: https://neos-server.org/neos/ to solve their Temoa models. If users do not have a local install of cplex, this is a potentially faster option. Note that when using NEOS, both --neos and --solver must be used in the config file.