Reformulate DemandActivityConstraintIndices
The prior version of this function was obsfucated and took a long time to construct. This new implementation is streamlined and much faster. With this new implementation of the DemandActivityConstraintIndices, we also need to make sure to skip time slices (s, d) for which a given demand is not defined.